Extracted generate stocklist text function
